requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
openai==1.25.1
httpx==0.27.0
python-dotenv==1.0.1
pydub==0.25.1
SpeechRecognition==3.10.1
Pillow==10.3.0
